Move errors package under commands
