+++ commands/scale_cluster_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/Jeffail/gabs"
 
 	"github.com/giantswarm/gsctl/config"
-	"github.com/giantswarm/gsctl/errors"
+	"github.com/giantswarm/gsctl/commands/errors"
 	"github.com/giantswarm/gsctl/flags"
 )
