+++ commands/show/cluster/command_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/giantswarm/gsctl/config"
-	"github.com/giantswarm/gsctl/errors"
+	"github.com/giantswarm/gsctl/commands/errors"
 	"github.com/giantswarm/gsctl/flags"
 )
